feat(api/applications): query calculations moved to .service (BOAS-92)
